Сделал класс GameView, вынес туда методы, общие для GameArtis & GameGenre
